+++ src/index.js
@@ -3,7 +3,7 @@
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
 
-import { create, seed } from './handlers/index.js'
+import { create, populate } from './handlers/index.js'
 
 yargs(hideBin(process.argv))
   .scriptName('strabot')
@@ -26,5 +26,5 @@ yargs(hideBin(process.argv))
     description: 'The platform the bot will run it',
     type: 'string'
   })
-  .command('seed', 'Creates initial data on Manager', seed)
+  .command('populate', 'Creates example data on manager', populate)
   .parse()
